Created user_request blueprint and updated remove
