+++ app/user_request/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+user_request = Blueprint('user_request', __name__)
+
+from . import views
